bot_telegram/requirements.txt
# Bot de Telegram para Ganadería San Juan
python-telegram-bot==20.7
python-dotenv==1.0.0
apscheduler==3.10.4
flask==3.0.0
pytz==2024.1
